requirements.txt
Flask==3.0.3
flask-cors==6.0.2
gspread==6.1.0
google-auth==2.29.0
google-generativeai==0.8.6
gunicorn==25.3.0
